fix: code cleanUp, move cors allowed to const
